Kernel: Basic screen terminal (WIP, currently signed integers do not display).

+++ kernel/types.hpp
@@ -5,7 +5,7 @@ using byte = unsigned char;
 using u32  = unsigned int;
 using u64  = unsigned long;
 using i32  = int;
-using i64  = long;
+using i64  = long long;
 using f32  = float;
 using f64  = double;
